requirements.txt
jieba==0.42.1
wordcloud==1.9.2
Pillow==10.1.0 